% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.table.R
\name{as.matrix.table}
\alias{as.matrix.table}
\title{Table to Matrix}
\usage{
\method{as.matrix}{table}(x)
}
\arguments{
\item{x}{A table (see \code{\link[base]{table}}).}
}
\value{
Logical value(s).
}
\description{
Convert a table object to a matrix.
}
\examples{
# Vector:
x <- table(rpois(100, 5))
as.matrix(x)

# Two dimensional contigency table:
x <- table(data.frame(a = rpois(100, 5), b = rpois(100, 10)))
as.matrix(x)

}
