% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.R
\name{clear}
\alias{clear}
\alias{clg}
\alias{clm}
\alias{clc}
\title{Clear}
\usage{
clg(x)

clm()

clc()
}
\arguments{
\item{x}{Graphic handle(s)}
}
\description{
Functions to clear or erase items.
}
\section{Functions}{

\describe{
   \item{\code{clg}}{Clear graphics windows.}
   \item{\code{clm}}{Clear all variables from memory.}
   \item{\code{clc}}{Clear the R console.}
}
}

\examples{
\dontrun{
# Clear graphics:
clg()
clg(grDevices::dev.list())

# Clear variables:
clm()

# Clear R console:
clc()
}

}
