% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflate.R
\name{collapse.data.frame}
\alias{collapse.data.frame}
\title{Collapse singluar fields
Expand data frame so its fields contain singluar attributes.
- Including header information?
- Do not necessarily remove attributes.}
\usage{
\method{collapse}{data.frame}()
}
\description{
Collapse singluar fields
Expand data frame so its fields contain singluar attributes.
- Including header information?
- Do not necessarily remove attributes.
}
