% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress.R
\name{compress}
\alias{compress}
\alias{compress.data.frame}
\alias{expand.data.frame}
\title{Compress or Expand Data Frames}
\usage{
compress(x, ...)

\method{compress}{data.frame}(x, keep, na.rm = TRUE, remove = "", unique = TRUE, ...)

\method{expand}{data.frame}(x, keep, ...)
}
\arguments{
\item{x}{Data object.}
}
\description{
Function to remove or restore redundant or empty elements from a data object.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Remove empty data columns.

\item \code{data.frame}: Attach data hidden as attributes to a data frame.
}}

\examples{
x <- data.frame(year = 2018, count = rpois(10, 5), survey = "scs", group = rep(1:2, each = 5))
x <- compress(x)
print(x)
attributes(x)
attr(x, "status") <- "critical"
expand(x)
attr(x, "units") <- c(length = "millimeters", weight = "grams")
expand(x)
}
