% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{date}
\alias{date}
\alias{date.default}
\alias{date.character}
\alias{date.data.frame}
\title{Date Extraction and Conversion}
\usage{
date(x, ...)

\method{date}{default}(x, year, month, day, ...)

\method{date}{character}(x, ...)

\method{date}{data.frame}(x, year, month, day, ...)
}
\arguments{
\item{x}{Object.}

\item{year}{Date year.}

\item{month}{Date month.}

\item{day}{Date day.}
}
\value{
Generally a \code{POSIXct} or \code{POSIXt} object.
}
\description{
These are convenience functions to extract a date field from an object, such as a character
string or a data frame.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default date method.

\item \code{character}: Convert character string to date.

\item \code{data.frame}: Extract date from a data frame.
}}

\examples{
date() # Base method.

# Single day:
date(year = 2000, month = 9, day = 19)

# Entire month:
date(year = 2000, month = 9, day = 1:30)

}
