% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deblank.R
\name{deblank}
\alias{deblank}
\title{Remove Leading and Trailing Spaces}
\usage{
deblank(x)
}
\arguments{
\item{x}{Character string(s).}
}
\description{
Removes leading and trailing spaces from character strings.
}
\examples{
x <- "  Why hello there. "
print(deblank(x))

}
