% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflate.R
\name{deflate}
\alias{deflate}
\title{Deflate or Inflate}
\usage{
deflate(x, ...)
}
\arguments{
\item{x}{Object, usually some type of data frame.}
}
\description{
Inflate and deflate objects.
}
\section{Functions}{

\describe{
   \item{\code{deflate}}{Generic \code{deflate} method.}
   \item{\code{inflate}}{Generic \code{inflate} method.}
   \item{\code{deflate.data.frame}}{Collapse columns with only single unique entries and store as object attributes.}
   \item{\code{inflate.data.frame}}{Expand object attributes and store as data columns.}
}
}

\examples{

x <- data.frame(year = 2018, count = rpois(10, 5), survey = "scs", group = rep(1:2, each = 5))
print(x)
x <- deflate(x)
print(x)
attributes(x)
x <- inflate(x)
print(x)

}
