% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\alias{describe.NULL}
\alias{describe.logical}
\alias{describe.numeric}
\alias{describe.character}
\alias{describe.factor}
\alias{describe.data.frame}
\title{Describe Contents}
\usage{
describe(x, ...)

\method{describe}{`NULL`}(x, ...)

\method{describe}{logical}(x, ...)

\method{describe}{numeric}(x, digits = 3, max = 10, sep = "-", ...)

\method{describe}{character}(x, max = 10, sep = "', '", ...)

\method{describe}{factor}(x, max = 10, sep = ", ", ...)

\method{describe}{data.frame}(x, indent = 3, drop = FALSE, ...)
}
\arguments{
\item{x}{Data frame.}

\item{...}{Other arguments (not used).}

\item{digits}{Number of significant digits to display for real numbers (default = 3).}

\item{max}{Maximum number of different elements to display (default = 10).}

\item{sep}{Character string specifying the separator to be used in concatenating unique values.}

\item{indent}{Number of spaces to indent field variables when displaying a data frame (default = 3).}

\item{drop}{Logical value specifying whether to ignore empty variables in the output.}
}
\description{
These functions provide a summary description of the contents of different types of R objects.
Unique elements are shown and sorted so that the minimum and maximum values are visible. For
integer values, ranges are shown. For data frames, the total number of NA or missing
values are also shown, along with the contents of each variable field.
}
\section{Methods (by class)}{
\itemize{
\item \code{describe(`NULL`)}: NULL method.

\item \code{describe(logical)}: Summary contents of a logical vector.

\item \code{describe(numeric)}: Summary contents of a numeric vector.

\item \code{describe(character)}: Summary contents of a character vector.

\item \code{describe(factor)}: Summary contents of a factor.

\item \code{describe(data.frame)}: Summary contents of a data frame.

}}
\examples{
# Integer examples:
describe(1:10)
describe(seq(2, 10, by = 2))
describe(seq(2, 40, by = 2))

# Decimal numeric examples:
describe(rnorm(10))
describe(rnorm(20))

# Character string examples:
describe(LETTERS[1:5])
describe(LETTERS)

# Create data frame
x <- data.frame(categories = LETTERS[1:5],
                n = rpois(5, lambda=1),
                values = rnorm(5))
describe(x)

# Bigger example:
x <- data.frame(categories = factor(sample(LETTERS[1:5], 25, replace = TRUE),
               levels = LETTERS[1:5]),
               n = rpois(25, lambda = 10),
               values = sample(c(rnorm(20), rep(NA, 5))),
               result = runif(25) > 0.3,
               missing = NA,
               stringsAsFactors = FALSE)
describe(x)

}
