% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-set.R
\name{description<-}
\alias{description<-}
\alias{description<-.default}
\alias{description}
\alias{description.default}
\title{Assign a Description}
\usage{
description(x, ...) <- value

\method{description}{default}(x, y, ...) <- value

description(x, ...)

\method{description}{default}(x, ...)
}
\arguments{
\item{x}{Target object.}

\item{...}{Other arguments (not used).}

\item{y}{Character string(s) specifying the name(s) of the variables or attributes to be assigned a description string.}

\item{value}{Character string(s) specifying the text description to be assigned.}
}
\value{
An object with a \code{description} attribute attached to it.
}
\description{
Assigns a text description to an object or to other attributes.
}
\details{
The \code{description} attribute may be erased by assigning a \code{NULL} value to it.
}
\section{Functions}{
\itemize{
\item \code{description<-.default}: Default 'description' assignment method.

\item \code{description}: Generic 'description' extraction method.

\item \code{description.default}: Default 'description' extraction method.
}}

\examples{
x <- data.frame(year = 2010:2014, measurement = rnorm(5))
description(x) <- "A simple table"
description(x, "year") <- "Year the sample was taken."
description(x, "measurement") <- "Length of the specimen."

# ... or equivalently:
description(x, c("year", "measurement")) <- c("Year the sample was taken.",
  "Length of the specimen.")
description(x) <- c(year = "Year the sample was taken.",
  measurement = "Length of the specimen.")
description(x) <- list(year = "Year the sample was taken.",
  measurement = "Length of the specimen.")

# Show attributes:
attributes(x)

# Erase 'description' attribute:
description(x, "year") <- NULL  # Remove 'year' description.
description(x) <- NULL          # Remove all descriptions.

}
\seealso{
\code{\link{metadata}}, \code{\link{key}}, \code{\link{units}}, \code{\link{fmt}}
}
