% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.github.R
\name{dir.github}
\alias{dir.github}
\title{List Git Repository Files}
\usage{
dir.github(username, repository)
}
\arguments{
\item{username}{GitHub user name.}

\item{repository}{GitHub repository name.}
}
\value{
Character vector of files and directories.
}
\description{
This function lists available files and directories on a GitHub repository.
}
