% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.r
\name{excel}
\alias{excel}
\alias{excel.default}
\alias{excel.data.frame}
\title{Export data frame to Excel}
\usage{
excel(x, ...)

\method{excel}{default}(x, ...)

\method{excel}{data.frame}(x, header = TRUE, row.names = FALSE, ...)
}
\arguments{
\item{x}{Data frame.}

\item{header}{Logical value specifying whether data frame column names should be exported.}

\item{row.names}{Logical value specifying whether data frame row names should be exported.}

\item{Further}{arguments passed onto the \code{\link[utils]{write.table}} function.}
}
\description{
This function dumps a data frame into a temporary tab-delimited table and opens an Excel session for the file.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default key method.

\item \code{data.frame}: Export data frame.
}}

\examples{

# Send a vector to Excel:
excel(1:10)

# Send a 25 x 4 matrix of random numbers to Excel, with no headers:
x <- matrix(runif(100), ncol = 4)
excel(x, header = FALSE)

}
