% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.locate.R
\name{file.locate}
\alias{file.locate}
\title{Find Package Files}
\usage{
file.locate(pattern, path, package, ...)
}
\arguments{
\item{pattern}{Character string(s) specifying the patterns to search for in the file name.}

\item{path}{Character string specifying the file path to search.}

\item{package}{Character string specifying the package name.}

\item{full.names}{Whether}
}
\value{
Character vector.
}
\description{
Searches for the path and name of file.
}
\examples{
file.locate(package = "gulf.data", pattern = "species.csv")
file.locate(package = "gulf.data", pattern = "*.csv")

}
