% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt-set.R
\name{fmt<-}
\alias{fmt<-}
\alias{fmt<-.default}
\alias{fmt}
\alias{fmt.default}
\title{Assign Data Format}
\usage{
fmt(x, ...) <- value

\method{fmt}{default}(x, y, ...) <- value

fmt(x, ...)

\method{fmt}{default}(x, ...)
}
\arguments{
\item{x}{Target object.}

\item{...}{Other arguments (not used).}

\item{y}{Character string(s) specifying the name(s) of the variables or attributes to be assigned data formats.}

\item{value}{Character string(s) specifying the data format to be assigned.}
}
\value{
An object with a \code{fmt} attribute attached to it.
}
\description{
Assigns a data format(s) to an object.
}
\details{
The \code{fmt} attribute may be erased by assigning a \code{NULL} value to it.
}
\section{Functions}{
\itemize{
\item \code{fmt<-.default}: Default 'fmt' assignment method.

\item \code{fmt}: Generic 'fmt' extraction method.

\item \code{fmt.default}: Default 'fmt' extraction method.
}}

\examples{
x <- data.frame(year = 2010:2014, month = 6, day = 1:5, measurement = rnorm(5))
fmt(x, "year")  <- "YYYY"
fmt(x, "month") <- "MM"
fmt(x, "day")   <- "DD"

# Show attributes:
attributes(x)

# Erase 'fmt' attribute:
fmt(x) <- NULL

}
\seealso{
\code{\link{metadata}}, \code{\link{key}}, \code{\link{description}}, \code{\link{units}}
}
