% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gulf.utils-package.R
\docType{package}
\name{gulf.utils-package}
\alias{gulf.utils}
\alias{gulf.utils-package}
\title{gulf.utils: Southern Gulf of Saint Lawrence Utility Functions}
\description{
Set of utility functions used to develop the 'gulf' family of packages.
}
\author{
\strong{Maintainer}: Tobie Surette \email{Tobie.Surette@dfo-mpo.gc.ca}

}
\keyword{internal}
