% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gulf.utils.R
\docType{package}
\name{gulf.utils}
\alias{gulf.utils-package}
\alias{gulf.utils}
\title{\strong{gulf.utils} Utility Functions}
\description{
Package of basic utility functions for use in the southern Gulf of Saint Lawrence fisheries' science.
}
\section{Functions and Methods}{

\describe{
   \item{\code{\link{clear}}}{Clearing functions.}
   \item{\code{\link{compress}}}{Remove empty data from object(s).}
   \item{\code{\link{date}}}{Date formatting functions.}
   \item{\code{\link{deblank}}}{Remove leading, trailing and repeating blanks in character strings.}
   \item{\code{\link{describe}}}{Return descriptive data summaries.}
   \item{\code{\link{excel}}}{Output data in MS Excel.}
   \item{\code{\link{import}}}{Import data between data frames.}
   \item{\code{\link{install.gulf}}}{Install suite of gulf packages.}
   \item{\code{\link{is.installed}}}{Check if R package is installed.}
   \item{\code{\link{language}}}{Format or set package languages.}
   \item{\code{\link{locate}}}{Search for data files.}
   \item{\code{\link{match}}}{Find matching rows between data frames.}
   \item{\code{\link{matrix2xyz}}}{Matrix to XYZ format conversion functions.}
   \item{\code{\link{na.rm}}}{Remove NA elements.}
   \item{\code{\link{pretty.log}}}{Returns a pretty log-scale which spans numeric data.}
   \item{\code{\link{repvec}}}{Repeat vectors along matrix dimensions.}
   \item{\code{\link{sort}}}{Sort data frame.}
   \item{\code{\link{spelling}}}{Correct common spelling mistakes.}
   \item{\code{\link{tictoc}}}{Chronometer functions.}
   \item{\code{\link{time}}}{Time formatting functions.}
   \item{\code{\link{time2unit}}}{Time unit conversion functions.}
   \item{\code{\link{translate}}}{Translate common fisheries science terms.}
   \item{\code{\link{unaccent}}}{Remove french accents from character strings.}
   \item{\code{\link{words}}}{Return word content in text or data.}
}
}

\author{
\strong{Maintainer}: Tobie Surette \email{Tobie.Surette@dfo-mpo.gc.ca}

}
