% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\alias{import<-}
\alias{import<-.data.frame}
\title{Import Data}
\usage{
import(x, ...) <- value

\method{import}{data.frame}(x, variables, fill.na, group, ...) <- value
}
\arguments{
\item{x}{Target object.}

\item{...}{Other parameters (not used).}

\item{value}{Object containing catch data to be assigned.}

\item{variables}{Character string(s) specifying the fields to import from the source object.}

\item{fill.na}{Value to be used to fill in target field entries with no matching source entries.}

\item{group}{Grouping variable.}
}
\description{
Import from one object into another.
}
\examples{
\dontrun{
library(gulf.data)

# Import snow crab survey by-catch data:
x <- read.scsset(year = 2020, valid = 1)
y <- read.scscat(year = 2020, species = c("cod", "plaice"))
y$species <- species(y$species)
import(x, fill = 0, var = "number.caught", group = "species") <- y

# Import snow crab category data:
x <- read.scsset(year = 2020, valid = 1)
b <- read.scsbio(2020)
import(x, fill = 0) <- catch(b, category = c("M", "COM", "MF"))
print(x$M)
}
}
\seealso{
\code{\link[gulf.utils]{match}}
}
