% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.installed.R
\name{is.installed}
\alias{is.installed}
\title{Check Package Installation}
\usage{
is.installed(package.name)
}
\arguments{
\item{package.name}{Name(s) of the packages to be checked.}
}
\value{
Logical value(s).
}
\description{
This function checks whether an R package(s) is installed.
}
\examples{

# Check whether there is a 'base' and 'splines' package:
is.installed(c("base", "splines"))

}
