% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.key.R
\name{is.key}
\alias{is.key}
\alias{is.key.default}
\alias{is.key.data.frame}
\title{Index Key Test}
\usage{
is.key(x, ...)

\method{is.key}{default}(x, key, ...)

\method{is.key}{data.frame}(x, key, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Other arguments (not used).}

\item{key}{Character string specifying the fields consituting an index key for an object.}
}
\description{
Checks whether an index key exists or is valid.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default \code{is.key} function.

\item \code{data.frame}: Check whether key is a valid index key for a data frame.
}}

\examples{
# Build sample data:
x <- data.frame(year = 1990:2014, measurement = rpois(25))
key(x) <- "year" # Assign key.

is.key(x, "measurement") # Generally FALSE
is.key(x, "year")
is.key(x)  # No need to specify 'year'.

}
\seealso{
\code{\link{key}}
}
