% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_-.R
\name{key<-}
\alias{key<-}
\title{Assign Index Key}
\usage{
key(x, ...) <- value
}
\arguments{
\item{x}{Object.}

\item{...}{Other arguments (not used).}

\item{value}{Index key to be assigned.}
}
\description{
Assign the variable name(s) which make up and index key for a given object.
}
\examples{
x <- data.frame(year = 2010:2014, measurement = rnorm(5))
key(x) <- "year" # Assign key.
key(x) # Retrieve key.

}
