% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{key}
\alias{key}
\alias{key.default}
\title{Retrieve Index Key}
\usage{
key(x, ...)

\method{key}{default}(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Other arguments (not used).}

\item{value}{Index key to be assigned.}
}
\description{
Retrieve the variable name(s) which make up and index key for a given object.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default key method.
}}

\examples{
x <- data.frame(year = 2010:2014, measurement = rnorm(5))
key(x) <- "year" # Assign key.
key(x) # Retrieve key.

}
