% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{language}
\alias{language}
\title{Language Options}
\usage{
language(x)
}
\arguments{
\item{x}{Character strings.}
}
\description{
Function to standardize language options.
}
\examples{
language()

language("f")     # Returns 'french'.
language("franc") # Returns 'french'.
language("ang")   # Returns 'english'.
language("lat")   # Returns 'latin'.

# Messy inputs:
language(c("en", "ENgli", "fr", "francais", "fr", "lat"))
}
