% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{lexicon}
\alias{lexicon}
\alias{lexicon.default}
\alias{lexicon.character}
\title{Build Lexicon}
\usage{
lexicon(x, ...)

\method{lexicon}{default}(x, file)

\method{lexicon}{character}(x)
}
\arguments{
\item{x}{Character vector or an object coercible to a text objexct.}

\item{file}{Character vector specifying a file to be read.}
}
\description{
Returns the different words used in a an object with text.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Returns a character vector of component words.

\item \code{character}: Returns a character vector of component words.
}}

