% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{locate}
\alias{locate}
\alias{locate.default}
\title{Find Data Files}
\usage{
locate(x, ...)

\method{locate}{default}(x, file, path, package, keywords, full.names = TRUE, ...)
}
\arguments{
\item{file}{Character string(s) specifying the files, including regular expressions, to search for in the file name.}

\item{path}{Character string specifying the file path to search.}

\item{package}{Character string specifying the package name.}

\item{full.names}{Logical value specifying whether to return the path names along with the file names.}
}
\value{
Character vector.
}
\description{
Searches for the path(s) and name of package data file(s).
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default function to locate data files.
}}

\examples{
locate(package = "gulf.data", file = "species.csv")
locate(package = "gulf.data", file = "*.csv")

}
\seealso{
\code{\link[base]{grep}}
}
