% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match}
\alias{match}
\alias{match.default}
\alias{match.data.frame}
\title{Match Data Frame Rows}
\usage{
match(x, ...)

\method{match}{default}(x, ...)

\method{match}{data.frame}(x, y, by, key, ...)
}
\arguments{
\item{x}{Source data frame whose elements will be searched for in the target data frame \code{y}.}

\item{...}{Further arguments passed onto \code{\link[base]{match}}.}

\item{y}{Target data frame.}

\item{by}{Column name(s) which will be used for matching elements.
The default is the common column names for \code{x} and \code{y}.}

\item{key}{Character vector defining index key.}
}
\description{
Returns the row indices of a target data frame which match those of a source data frame.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default match method, see \code{\link[base]{match}}.

\item \code{data.frame}: Match rows between two data frames.
}}

\examples{
# Create a simple data frame:
x <- data.frame(
        u = c(2, 2, 4),
        v = c(1, 2, 3))

y <- data.frame(
        u = c(1, 1, 1, 1, 2, 2, 2, 3),
        v = c(1, 2, 3, 4, 1, 2, 3, 3))

# Get the list of indices matching the rows of 'x' to 'y':
index <- match(x, y)

}
