% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix2xyz.R
\name{matrix2xyz}
\alias{matrix2xyz}
\alias{xyz2matrix}
\title{Matrix Format Conversion}
\usage{
matrix2xyz(m, na.rm = FALSE)

xyz2matrix(x, y, z, fun)
}
\arguments{
\item{m}{Matrix.}

\item{na.rm}{Logical value specifying whether to remove \code{NA} values in \code{z} the output.
The default is \code{FALSE}.}

\item{x}{Vector containing matrix row indices. Alternatively, \code{x} may be a
three-column matrix or data frame containing columns specifying
\code{x}, \code{y} and \code{z}.}

\item{y}{Vector containing the matrix column indices. Must be the same length as \code{x}.}

\item{z}{Vector containing the matrix elements which are mapped by \code{x} and \code{y}.
Must be the same length as \code{x} and \code{y}.}

\item{fun}{Function (e.g. \code{\link[base]{mean}}) to be used on \code{z} when the indices \code{x}
and \code{y} have duplicate entries.}
}
\value{
Matrix
}
\description{
These are functions to convert between different matrix formats. Generally, we may convert a two-dimensional
array of values to a linear xyz format which contains the coordinates or indices of the rows and columns
corresponding to each matrix element \code{z}.
}
\section{Functions}{
\itemize{
\item \code{matrix2xyz}: Convert from matrix to xyz format.

\item \code{xyz2matrix}: Convert from xyz to matrix format.
}}

\examples{
# Define  random matrix:
m <- round(matrix(runif(40), nrow = 8), 2)

# Convert to XYZ format:
v <- matrix2xyz(m)

# Convert back to matrix format:
xyz2matrix(x = v[,"x"], y = v[,"y"], z = v[,"z"])

# Equivalently:
xyz2matrix(v)

# Convert XYZ elements with duplicate entries to matrix:
x <- c(1, 1, 3, 2, 3)
y <- c(2, 2, 1, 2, 3)
z <- 1:5
m <- xyz2matrix(x, y, z, mean)

# Convert back to XYZ format:
v <- matrix2xyz(m, na.rm = TRUE)

}
