% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\alias{metadata.data.frame}
\title{Extract Metadata}
\usage{
metadata(x, ...)

\method{metadata}{data.frame}(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Other arguments (not used).}
}
\description{
Function to extract metadata from an object.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Extract metadata for a data frame.
}}

\seealso{
\code{\link{key}}, \code{\link{description}}, \code{\link{units}}, \code{\link{fmt}}
}
