% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nword.R
\name{nword}
\alias{nword}
\title{Number of Words}
\usage{
nword(x, ...)
}
\arguments{
\item{x}{Character vector or an object coercible to a text objexct.}

\item{...}{Further arguments (not used).}
}
\description{
Return the number of words contained in a text.
}
\examples{
nword("The punctuation, in this phrase; is ... really: 'terrible'.")
}
