% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty.log.R
\name{pretty.log}
\alias{pretty.log}
\title{Pretty sequence of logarithmic values}
\usage{
\method{pretty}{log}(x, base, n = 5, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{An integer giving the target number of intervals. Non-integer values are rounded down.}

\item{...}{Further arguments passed on to \code{\link[base]{pretty}}.}
}
\description{
Returns a sequence of values on the logarithmic scale suitable for labelling on a logarithmic plot, i.e.
the values are equally spaced on the log-scale.
}
\examples{
# Exponentail breakpoints across values from 0 to 10:
pretty.log(0:10)

# Generate exponential breakpoints for random numbers:
pretty.log(runif(100)*100)
}
