% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-set.R
\name{project<-}
\alias{project<-}
\alias{project<-.default}
\alias{project}
\alias{project.default}
\title{Assign a Project Identifier}
\usage{
project(x, ...) <- value

\method{project}{default}(x, ...) <- value

project(x, ...)

\method{project}{default}(x, ...)
}
\arguments{
\item{x}{Target object.}

\item{...}{Other arguments (not used).}

\item{value}{Character string(s) specifying the project identifier to be assigned.}
}
\value{
An object with a \code{project} attribute attached to it.
}
\description{
Assigns a project identifier to an object.
}
\details{
The \code{project} attribute may be erased by assigning a \code{NULL} value to it.
}
\section{Functions}{
\itemize{
\item \code{project<-.default}: Default 'project' assignment method.

\item \code{project}: Generic 'project' extraction method.

\item \code{project.default}: Default 'project' extraction method.
}}

\examples{
x <- data.frame(year = 2010:2014, measurement = rnorm(5))
project(x) <- "test.survey"

# Show attributes:
attributes(x)

# Erase 'project' attribute:
project(x) <- NULL          # Remove all projects.

}
\seealso{
\code{\link{metadata}}, \code{\link{key}}, \code{\link{description}}, \code{\link{units}}, \code{\link{fmt}}
}
