% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\title{Fix Encoding}
\usage{
recode(x, ...)
}
\arguments{
\item{x}{R object containing character data.}

\item{...}{Not used.}
}
\value{
R object with standardized character encoding.
}
\description{
Fix encoding issues.
}
