% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repvec.R
\name{repvec}
\alias{repvec}
\title{Repeat vectors}
\usage{
repvec(x, n = 1, dim = NULL, nrow = NULL, ncol = NULL)
}
\arguments{
\item{x}{Scalar or vector.}

\item{n}{Integer specifying the number of times \code{x} is to be repeated.}

\item{dim}{Integer specifying along which dimension \code{x} is to be repeated.}

\item{nrow}{Integer specifying the number of row-wise repetitions of \code{x}.}

\item{ncol}{Integer specifying the number of column-wise repetitions of \code{x}.}
}
\description{
Create matrices by repeating vectors row-wise or column-wise a specified number of times.
}
\examples{
  repvec(1, 12)              # Column vector of ones.
  repvec(1, ncol = 12)       # Row vector of ones.
  repvec(1:10, 12, dim = 2)  # 10x12 matrix with 1:10 in each column.
  repvec(1:10, 12, dim = 1)  # 12x10 matrix with 1:10 in each row.
  repvec(1:10, ncol = 12)    # 10x12 matrix with 1:10 in each column.

}
