% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.data.frame.R
\name{sort.data.frame}
\alias{sort.data.frame}
\title{Sort Data Frame}
\usage{
\method{sort}{data.frame}(x, by = NULL, increasing = TRUE)
}
\arguments{
\item{x}{Data frame.}

\item{by}{Data frame variable names or column indices by which to sort the data frame.}

\item{increasing}{Logical value specifying whether to sort by increasing order.}
}
\description{
Sorts a dta frame by one or more columns.
}
\examples{
# Create a simple data frame:
x <- data.frame(
        u = c(1, 1, 1, 1, 2, 2, 2, 3),
        v = c(1, 2, 3, 4, 1, 2, 3, 3))

sort(x, by = "u")
sort(x, by = "v")
sort(x, by = c("u", "v"))  # Sort by both variables.

}
