% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spelling.R
\name{spelling}
\alias{spelling}
\title{Correct Spelling}
\usage{
spelling(x)
}
\arguments{
\item{x}{Character string(s).}
}
\description{
Functions for correcting common fishery and biological spelling errors.
}
\examples{
spelling("aligatore")
spelling("yelow  tail")
}
