% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squeeze.R
\name{squeeze}
\alias{squeeze}
\alias{squeeze.data.frame}
\title{Remove Redundant Elements}
\usage{
squeeze(x, ...)

\method{squeeze}{data.frame}(x, ...)
}
\arguments{
\item{x}{Data object.}
}
\description{
Function to remove redundant elements from a data object.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Remove empty data columns.
}}

