% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.list.R
\name{subset.list}
\alias{subset.list}
\title{List Subset Function}
\usage{
\method{subset}{list}(x, index, ...)
}
\arguments{
\item{x}{A list vector.}

\item{index}{Regular list index.}

\item{...}{Named arguments which refer to list cell contents.}
}
\value{
A list object is returned.
}
\description{
Returns a subset of a list object corresponding to a set of specified
             membership criteria.
}
\seealso{
\code{\link[base]{subset}}
}
