% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tic.R
\name{tic}
\alias{tic}
\alias{toc}
\title{Calculate elapsed time}
\usage{
tic(gcFirst = TRUE)
}
\arguments{
\item{gcFirst}{Logical value.}
}
\description{
Calculate elapsed time
}
\examples{
tic()
   for (i in 1:1000000) x <- i
toc()

}
