% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tictoc.R
\name{tictoc}
\alias{tictoc}
\alias{tic}
\alias{toc}
\title{Calculate elapsed time}
\usage{
tic(gcFirst = TRUE)

toc()
}
\arguments{
\item{gcFirst}{Logical value.}
}
\description{
Calculates elapsed time between calls to \code{tic} and \code{toc}.
}
\section{Functions}{

\describe{
   \item{\code{tic}}{Start timer.}
   \item{\code{toc}}{Stop timer and report elapsed time.}
}
}

\examples{
\dontrun{
tic()
   for (i in 1:1000000) x <- i
toc()
}
}
