% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{time}
\alias{time}
\alias{time.default}
\alias{time.logical}
\alias{time.numeric}
\alias{time.character}
\alias{time.data.frame}
\title{Time Extraction and Conversion}
\usage{
time(x, ...)

\method{time}{default}(x, date, time, year, month, day, hour = 0, minute = 0, second = 0, ...)

\method{time}{logical}(x)

\method{time}{numeric}(x)

\method{time}{character}(x)

\method{time}{data.frame}(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Further arguments passed onto}

\item{date}{Date object or formatted character string representing a date.}

\item{time}{Time object or formatted character string representing a time.}

\item{year}{Date year.}

\item{month}{Date month.}

\item{day}{Date day.}

\item{hour}{Time hour.}

\item{minute}{Time minute.}

\item{second}{Time second.}
}
\value{
Generally a \code{POSIXct} or \code{POSIXt} object.
}
\description{
These are convenience functions to extract a time field from an object, such as a character
string or a data frame.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default time method.

\item \code{logical}: Convert time from logical value.

\item \code{numeric}: Convert time from numeric value.

\item \code{character}: Convert time from character string.

\item \code{data.frame}: Extract time from data frame.
}}

\examples{
time() # Current time (base method).

# Equivalent nine o'clock specification:
time(9)
time("9")
time("9h")
time("9h00")
time("9:00")
time("9:00:00")

# All equivalent time strings:
time("194501")
time("19:45:01")
time("19h45m01")
time("19h45m01")

}
