% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time2unit.R
\name{time2unit}
\alias{time2unit}
\alias{time2sec}
\alias{time2min}
\alias{time2hour}
\alias{time2day}
\alias{julian}
\alias{week}
\title{Relative Time Conversion}
\usage{
time2unit(x, reference, units, ...)

time2sec(x, ...)

time2min(x, ...)

time2hour(x, ...)

time2day(x, ...)

julian(x)

week(x, reference = "Sunday")
}
\arguments{
\item{x}{Time variable.}

\item{reference}{Reference date, time or day of the week.}

\item{units}{Time units: \code{secs}, \code{mins}, \code{hours} or \code{days}.}
}
\description{
Calculate time difference relative to a specified reference time.
}
\section{Functions}{
\itemize{
\item \code{time2sec}: Calculate time difference in \bold{seconds}.

\item \code{time2min}: Calculate time difference in \bold{minutes}.

\item \code{time2hour}: Calculate time difference in \bold{hours}.

\item \code{time2day}: Calculate time difference in \bold{days}.

\item \code{julian}: Calculate Julian day, i.e. the day of the year relative to January 1st.

\item \code{week}: Calculate to in which \bold{week} of the year a date fall into.
}}

\seealso{
\code{\link[base]{difftime}}
}
