% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{en2fr}
\alias{en2fr}
\alias{fr2en}
\title{Translate Words}
\usage{
en2fr(x, plural = FALSE, feminine = FALSE)

fr2en(x, plural = FALSE, feminine = FALSE)
}
\arguments{
\item{x}{Character string(s).}

\item{plural}{Logical value specifying whether the plural form of a word is to be imposed.}

\item{feminine}{Logical value specifying whether the feminine form of a word is to be imposed.}
}
\description{
Functions for translating common fishery and biological words and terms. These
             functions are intended for translating words and common short phrases only.
}
\section{Functions}{
\itemize{
\item \code{en2fr}: Translate from english to french.

\item \code{fr2en}: Translate from english to french.
}}

\examples{
en2fr("male")
en2fr("crab")
en2fr("black")
en2fr("missing legs")
en2fr("fishing grids")
en2fr("survey")
}
