% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unaccent.R
\name{unaccent}
\alias{unaccent}
\title{Remove Accented Characters}
\usage{
unaccent(x)
}
\arguments{
\item{x}{Character string(s) where accented characters will be replaced by their unaccented versions.}
}
\description{
Replace french characters with unaccented versions.
}
