% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units-set.R
\name{units<-}
\alias{units<-}
\alias{units<-.default}
\alias{units}
\alias{units.default}
\title{Assign Observational Units}
\usage{
units(x, ...) <- value

\method{units}{default}(x, y, ...) <- value

units(x, ...)

\method{units}{default}(x, ...)
}
\arguments{
\item{x}{Target object.}

\item{...}{Other arguments (not used).}

\item{y}{Character string(s) specifying the name(s) of the variables or attributes to be assigned measurement units.}

\item{value}{Character string(s) specifying the units to be assigned.}
}
\value{
An object with a \code{units} attribute attached to it.
}
\description{
Assigns units to measured or observed data in an object.
}
\details{
The \code{units} attribute may be erased by assigning a \code{NULL} value to it.
}
\section{Functions}{
\itemize{
\item \code{units<-.default}: Default 'units' assignment method.

\item \code{units}: Generic 'units' extraction method.

\item \code{units.default}: Default 'units' extraction method.
}}

\examples{
x <- data.frame(year = 2010:2014, measurement = rnorm(5))
units(x, "measurement") <- "centimeters"

# Show attributes:
attributes(x)

# Erase 'units' attribute:
units(x) <- NULL

}
\seealso{
\code{\link{metadata}}, \code{\link{key}}, \code{\link{description}}, \code{\link{fmt}}
}
