% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{date}
\alias{date}
\alias{date.default}
\alias{date.character}
\alias{date.data.frame}
\alias{day.default}
\alias{month.default}
\alias{year.default}
\title{Date Extraction and Conversion}
\usage{
date(x, ...)

\method{date}{default}(x, year, month, day, ...)

\method{date}{character}(x, ...)

\method{date}{data.frame}(x, ...)

\method{day}{default}(x, ...)

\method{month}{default}(x, ...)

\method{year}{default}(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Further arguments.}

\item{year}{Date year.}

\item{month}{Date month.}

\item{day}{Date day.}
}
\value{
Generally a \code{POSIXct} or \code{POSIXt} object.
}
\description{
These are wrapper functions to extract a date field from an object, such as a character
string or a data frame.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default date method.

\item \code{character}: Convert character string to date.

\item \code{data.frame}: Extract date from a data frame.

\item \code{default}: Default method for extracting day from a date.

\item \code{default}: Default method for extracting month from a date.

\item \code{default}: Default method for extracting year field from a date.
}}

\examples{
date() # Base method.

# Single day:
date(year = 2000, month = 9, day = 19)

# Entire month:
date(year = 2000, month = 9, day = 1:30)

# Sorts out YYYYMMDD and DDMMYYYY:
date(c("2000-01-19", "19-01-2000", "2000/01/19", "00-1-19"))

# Apply to data frame:
x <- data.frame(Year = 2000, Month = 01, Day = 1:30)
date(x)

}
\seealso{
\code{\link[gulf.utils]{time}}
}
