% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Describe Data Frame}
\usage{
describe(x, ...)
}
\arguments{
\item{x}{Data frame.}

\item{...}{Other arguments (not used).}

\item{index}{Number of indent spaces (default = 3).}
}
\description{
This function provides a description of the unique and missing elements of a data frame on the R console.
Unique elements are sorted so that the minimum and maximum values are visible.
The total number of missing and NA values are also shown when the occur.
}
\examples{
# Create data frame
x <- data.frame(categories = LETTERS[1:5],
                n = rpois(5),
                values = rnorm(5)))

describe(x)

# Bigger example:
x <- data.frame(categories = LETTERS[1:25],
                n = rpois(25, lambda = 10),
                values = sample(c(rnorm(20), rep(NA, 5))),
                result = runif(25) > 0.3,
                missing = NA)

describe(x)
}
