% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.locate.R
\name{file.locate}
\alias{file.locate}
\title{Find Package Data Files}
\usage{
file.locate(pattern, path, package, full.names = TRUE, ...)
}
\arguments{
\item{pattern}{Character string(s) specifying the patterns to search for in the file name.}

\item{path}{Character string specifying the file path to search.}

\item{package}{Character string specifying the package name.}

\item{full.names}{Logical value specifying whether to return the path names along with the file names.}
}
\value{
Character vector.
}
\description{
Searches for the path(s) and name of package data file(s).
}
\examples{
file.locate(package = "gulf.data", pattern = "species.csv")
file.locate(package = "gulf.data", pattern = "*.csv")

}
