% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gulf.utils.R
\docType{package}
\name{gulf.utils}
\alias{gulf.utils}
\title{gulf.utils: sGSL Utility Functions}
\description{
A package of basic utility functions for use in the southern Gulf of Saint Lawrence
fisheries' science.
}
\section{\code{gulf.utils} functions}{


\code{\link{match}} matches the rows between data frames.
}

