% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key-set.R
\name{key<-}
\alias{key<-}
\alias{key<-.default}
\alias{key}
\alias{key.default}
\title{Assign or Retrieve an Index Key}
\usage{
key(x, ...) <- value

\method{key}{default}(x) <- value

key(x, ...)

\method{key}{default}(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Other arguments (not used).}

\item{value}{Character string(s) specifying the index key to be assigned.}
}
\description{
Assign or retrieve the variable name(s) which make up and index key for a given object.
}
\section{Functions}{
\itemize{
\item \code{key<-.default}: Default key assignment method.

\item \code{key}: Generic 'key' method.

\item \code{key.default}: Default 'key' method.
}}

\examples{
x <- data.frame(year = 2010:2014, measurement = rnorm(5))
key(x) <- "year" # Assign key.
key(x) # Retrieve key.

}
\seealso{
\code{\link{is.key}}, \code{\link{metadata}}, \code{\link{description}}, \code{\link{units}}, \code{\link{fmt}}
}
