% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sort.default}
\alias{sort.default}
\alias{sort.data.frame}
\title{Sort Data Frame}
\usage{
\method{sort}{default}(x, ...)

\method{sort}{data.frame}(x, by, increasing = TRUE, ...)
}
\arguments{
\item{x}{Data frame.}

\item{by}{Field variable names or column indices by which to sort a data frame.}

\item{increasing}{Logical value specifying whether to sort by increasing order.}
}
\description{
Sorts a data frame by one or more columns.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default sort method, see \code{\link[base]{sort}}.

\item \code{data.frame}: Sort a data frame.
}}

\examples{
# Create a simple data frame:
x <- data.frame(
        u = c(1, 1, 1, 1, 2, 2, 2, 3),
        v = c(1, 2, 3, 4, 1, 2, 3, 3))

sort(x, by = "u")
sort(x, by = "v")
sort(x, by = c("u", "v"))  # Sort by both variables.

}
\seealso{
\code{\link[base]{order}}
}
