% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{time}
\alias{time}
\alias{time.default}
\alias{time.character}
\alias{time.data.frame}
\title{Time Extraction and Conversion}
\usage{
time(x, ...)

\method{time}{default}(x, date, time, year, month, day, hour = 0, minute = 0, second = 0, ...)

\method{time}{character}(x)

\method{time}{data.frame}(x, ...)
}
\arguments{
\item{x}{Object.}

\item{year}{Date year.}

\item{month}{Date month.}

\item{day}{Date day.}
}
\value{
Generally a \code{POSIXct} or \code{POSIXt} object.
}
\description{
These are convenience functions to extract a time field from an object, such as a character
string or a data frame.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default time method.

\item \code{character}: Convert time from character string.

\item \code{data.frame}: Extract time from data frame.
}}

\examples{
time() # Base method.

# Single day:
date(year = 2000, month = 9, day = 19)

# Entire month:
date(year = 2000, month = 9, day = 1:30)

}
