% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time2unit.R
\name{time2unit}
\alias{time2unit}
\alias{time2sec}
\alias{time2min}
\alias{time2hour}
\alias{time2day}
\alias{julian}
\alias{week}
\title{Relative Time Conversion}
\usage{
time2unit(x, reference, units, ...)

time2sec(x, ...)

time2min(x, ...)

time2hour(x, ...)

time2day(x, ...)

julian(x)

week(x, reference = "Sunday")
}
\arguments{
\item{x}{Time variable.}

\item{reference}{Reference date, time or day of the week.}

\item{units}{Time units: \code{secs}, \code{mins}, \code{hours} or \code{days}.}

\item{...}{Not used.}
}
\description{
Calculate time difference relative to a specified reference time.
}
\section{Functions}{
\itemize{
\item \code{time2sec}: Calculate time difference in \bold{seconds}.

\item \code{time2min}: Calculate time difference in \bold{minutes}.

\item \code{time2hour}: Calculate time difference in \bold{hours}.

\item \code{time2day}: Calculate time difference in \bold{days}.

\item \code{julian}: Calculate Julian day, i.e. the day of the year relative to January 1st.

\item \code{week}: Calculate to in which \bold{week} of the year a date fall into.
}}

\examples{
# Create sample dates:
v <- date(c("2016-07-10", "2016-07-13", "2016-07-16", "2016-08-25", "2016-10-01"))

time2sec(v)   # Number of seconds relative to default Jan 1st, 1970.
time2sec(v, as.POSIXct("1900-01-01")) # Relative to default Jan 1st, 1900.
time2min(v, "1900-01-01")
time2hour(v, "1900-01-01")
time2day(v, "1900-01-01")

# Return the julian day for a set of dates.
julian(v)

}
\seealso{
\code{\link[base]{difftime}}
}
