% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words.R
\name{words}
\alias{words}
\alias{words.default}
\alias{words.character}
\title{Extract Words}
\usage{
words(x, ...)

\method{words}{default}(x, file, ...)

\method{words}{character}(x, ...)
}
\arguments{
\item{x}{Character vector or an object coercible to a text objexct.}

\item{...}{Further arguments (not used).}

\item{file}{Character vector specifying a file to be read.}
}
\description{
Return a list of words contained in a text.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Returns a character vector of component words.

\item \code{character}: Returns a character vector of component words.
}}

\examples{
words("The punctuation, in this phrase; is ... really: 'terrible'.")
}
